refactor: move inner class out of DefaultMessageStore: CommitLogDispatcherBuildConsumeQueue
